requirements.txt
fastapi
uvicorn
requests
supabase
python-dotenv
mangum
boto3
pytest
flake8